Move extracurriculars section into components folder
