+++ src/pages/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 
-import Extracurriculars from './extracurriculars';
-
 import Layout from '../components/Layout';
+
 import About from '../components/about';
 import Education from '../components/education';
 import Experience from '../components/experience';
 import Skills from '../components/skills';
 import Interests from '../components/interests';
+import Extracurriculars from '../components/extracurriculars';
 
 import Sidebar from '../components/Sidebar';
 import config from '../../config';
